+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - chore: Add ruby-lsp gem
 - refactor: Parse JSON with symbolize names
+- refactor: Make request body with symbolize keys
+- refactor: Make blocks with symbolize keys
 
 ## [v2.1.0] - 2024-01-19